Split DeferredLoadingContext out of World
